rename 'common' module to 'core'
