+++ backend/src/accounts/entities/storage-item.entity.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, ManyToOne, PrimaryColumn } from "typeorm";
 import { AccountEntity } from "./account.entity";
 import { AccountStorageDomain, AccountStorageItem } from "@flowser/shared";
-import { PollingEntity } from "../../common/entities/polling.entity";
+import { PollingEntity } from "../../core/entities/polling.entity";
 import {
   FlowAccountStorage,
   FlowAccountStorageDomain,
