+++ backend/src/app.module.ts
@@ -12,7 +12,7 @@ import { TransactionsModule } from "./transactions/transactions.module";
 import { EventsModule } from "./events/events.module";
 import { LogsModule } from "./logs/logs.module";
 import { FlowModule } from "./flow/flow.module";
-import { CommonModule } from "./common/common.module";
+import { CoreModule } from "./core/core.module";
 import { getDatabaseOptions } from "./database";
 
 @Global()
@@ -29,7 +29,7 @@ import { getDatabaseOptions } from "./database";
     EventsModule,
     LogsModule,
     FlowModule,
-    CommonModule,
+    CoreModule,
   ],
   controllers: [AppController],
 })
