+++ backend/src/core/core.module.ts
@@ -1,5 +1,5 @@
 import { Module } from "@nestjs/common";
-import { CommonService } from "./common.service";
+import { CommonService } from "./services/common.service";
 import { AccountsModule } from "../accounts/accounts.module";
 import { BlocksModule } from "../blocks/blocks.module";
 import { TransactionsModule } from "../transactions/transactions.module";
@@ -18,9 +18,9 @@ import { EventsModule } from "../events/events.module";
     EventsModule,
     LogsModule,
     TransactionsModule,
-    CommonModule,
+    CoreModule,
   ],
   providers: [CommonService],
   exports: [CommonService],
 })
-export class CommonModule {}
+export class CoreModule {}
