+++ backend/src/core/entities/polling.entity.ts
