+++ backend/src/core/interceptors/polling-response.interceptor.ts
