+++ backend/src/core/shutdown-handler.ts
