+++ backend/src/events/entities/event.entity.ts
@@ -1,4 +1,4 @@
-import { PollingEntity } from "../../common/entities/polling.entity";
+import { PollingEntity } from "../../core/entities/polling.entity";
 import { AfterLoad, Column, Entity, PrimaryColumn } from "typeorm";
 import { ExtendedFlowEvent } from "../../flow/services/aggregator.service";
 import { Event } from "@flowser/shared";
