+++ backend/src/events/events.controller.ts
@@ -7,7 +7,7 @@ import {
   Post,
 } from "@nestjs/common";
 import { EventsService } from "./events.service";
-import { PollingResponseInterceptor } from "../common/interceptors/polling-response.interceptor";
+import { PollingResponseInterceptor } from "../core/interceptors/polling-response.interceptor";
 import { ApiParam } from "@nestjs/swagger";
 import {
   GetAllEventsResponse,
