+++ backend/src/index.ts
@@ -1,8 +1,8 @@
 import { INestApplication, ValidationPipe } from "@nestjs/common";
 import { NestFactory } from "@nestjs/core";
 import { AppModule } from "./app.module";
-import { HttpExceptionFilter } from "./common/filters/http-exception.filter";
-import { Config, ConfigService } from "./common/config.service";
+import { HttpExceptionFilter } from "./core/filters/http-exception.filter";
+import { Config, ConfigService } from "./core/services/config.service";
 
 export { FlowEmulatorService } from "./flow/services/emulator.service";
 export { FlowCliService } from "./flow/services/cli.service";
