+++ backend/src/logs/logs.controller.ts
@@ -7,7 +7,7 @@ import {
   Post,
 } from "@nestjs/common";
 import { LogsService } from "./logs.service";
-import { PollingResponseInterceptor } from "../common/interceptors/polling-response.interceptor";
+import { PollingResponseInterceptor } from "../core/interceptors/polling-response.interceptor";
 import {
   GetAllLogsResponse,
   GetPollingLogsRequest,
