+++ backend/src/transactions/transactions.controller.ts
@@ -7,7 +7,7 @@ import {
   Post,
 } from "@nestjs/common";
 import { TransactionsService } from "./transactions.service";
-import { PollingResponseInterceptor } from "../common/interceptors/polling-response.interceptor";
+import { PollingResponseInterceptor } from "../core/interceptors/polling-response.interceptor";
 import { ApiParam, ApiQuery } from "@nestjs/swagger";
 
 import {
