+++ frontend/src/targets/electron/main.ts
@@ -63,7 +63,7 @@ app.on("ready", createWindow);
 
 // Quit when all windows are closed.
 app.on("window-all-closed", function () {
-  // On OS X it is common for applications and their menu bar
+  // On OS X it is core for applications and their menu bar
   // to stay active until the user quits explicitly with Cmd + Q
   if (process.platform !== "darwin") {
     app.quit();
@@ -71,7 +71,7 @@ app.on("window-all-closed", function () {
 });
 
 app.on("activate", function () {
-  // On OS X it's common to re-create a window in the app when the
+  // On OS X it's core to re-create a window in the app when the
   // dock icon is clicked and there are no other windows open.
   if (BrowserWindow.getAllWindows().length === 0) createWindow();
 });
