+++ shared/proto/entities/accounts.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 import "entities/transactions.proto";
-import "entities/common.proto";
+import "entities/core.proto";
 import "google/protobuf/struct.proto";
 
 message Account {
