+++ shared/proto/responses/accounts.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 import "entities/accounts.proto";
-import "responses/common.proto";
+import "responses/core.proto";
 
 message GetAllAccountsResponse {
   repeated Account accounts = 1;
