+++ shared/proto/responses/blocks.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 import "entities/blocks.proto";
-import "responses/common.proto";
+import "responses/core.proto";
 
 message GetAllBlocksResponse {
   repeated Block blocks = 1;
