+++ shared/proto/responses/projects.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 import "entities/projects.proto";
 import "entities/config.proto";
-import "responses/common.proto";
+import "responses/core.proto";
 
 message GetAllProjectsResponse {
   repeated Project projects = 1;
