+++ shared/proto/responses/transactions.proto
@@ -1,7 +1,7 @@
 syntax = "proto3";
 
 import "entities/transactions.proto";
-import "responses/common.proto";
+import "responses/core.proto";
 
 message GetAllTransactionsResponse {
   repeated Transaction transactions = 1;
